Add general header
