+++ docs/sitemap.xml
@@ -2,6 +2,6 @@
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
   <url>
     <loc>https://talks.tidy-finance.org/index.html</loc>
-    <lastmod>2025-04-10T10:10:28.996Z</lastmod>
+    <lastmod>2025-04-10T10:19:07.811Z</lastmod>
   </url>
 </urlset>
